continuous-compliance-engine/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
PyYAML==6.0.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
jinja2==3.1.5
